load vocab when training if cache found

+++ parser/commands/train.py
@@ -34,11 +34,19 @@ class Train(object):
         dev = Corpus.load(args.fdev)
         test = Corpus.load(args.ftest)
         embed = Embedding.load(args.fembed)
-        vocab = Vocab.from_corpus(corpus=train, min_freq=2)
-        vocab.read_embeddings(embed=embed, unk='unk')
-        torch.save(vocab, args.vocab)
-        FNULL = open(os.devnull, 'w')
-        subprocess.call(['gsutil', 'cp', args.vocab, args.cloud_address+args.vocab], stdout=FNULL, stderr=subprocess.STDOUT)
+
+        if not os.path.isfile(args.vocab):
+            FNULL = open(os.devnull, 'w')
+            subprocess.call(['gsutil', 'cp', args.cloud_address+args.vocab, args.vocab], stdout=FNULL, stderr=subprocess.STDOUT)
+        if not os.path.isfile(args.vocab):
+          vocab = Vocab.from_corpus(corpus=train, min_freq=2)
+          vocab.read_embeddings(embed=embed, unk='unk')
+          torch.save(vocab, args.vocab)
+          FNULL = open(os.devnull, 'w')
+          subprocess.call(['gsutil', 'cp', args.vocab, args.cloud_address+args.vocab], stdout=FNULL, stderr=subprocess.STDOUT)
+        else:
+          vocab = torch.load(args.vocab)
+       
         print(vocab)
 
         print("Load the dataset")
